minor style improvements, code refactor, remove PerksGrid file
